Additional cleanup (#188)
* moving test clients to one file (instead of being with SimulatedEVentHubs

* changing some parameter names in RateControlUtils for consistency

* changing from java getters/setters to scala-styled ones
