+++ core/src/main/scala/org/apache/spark/eventhubscommon/client/EventHubsOffsetTypes.scala
@@ -18,9 +18,7 @@
 package org.apache.spark.eventhubscommon.client
 
 private[spark] object EventHubsOffsetTypes extends Enumeration {
-
   type EventHubsOffsetType = Value
 
   val None, PreviousCheckpoint, InputByteOffset, EnqueueTime = Value
-
 }
