+++ core/src/test/scala/org/apache/spark/eventhubscommon/utils/SimulatedEventHubs.scala
@@ -19,9 +19,6 @@ package org.apache.spark.eventhubscommon.utils
 
 import com.microsoft.azure.eventhubs.EventData
 import org.apache.spark.eventhubscommon.EventHubNameAndPartition
-import org.apache.spark.eventhubscommon.client.EventHubsOffsetTypes.EventHubsOffsetType
-import org.apache.spark.eventhubscommon.client.{ Client, EventHubsOffsetTypes }
-import org.apache.spark.streaming.StreamingContext
 
 import scala.collection.mutable.ListBuffer
 
@@ -69,127 +66,5 @@ class SimulatedEventHubs(eventHubsNamespace: String,
         .groupBy(_._1)
         .map { case (k, v) => (k, v.flatMap(_._2).toArray) }
     messageStore = combinedData
-  }
-}
-
-// TODO: consolidate these to one mock object/companion
-class SimulatedEventHubsRestClient(eventHubs: SimulatedEventHubs)
-    extends Client
-    with TestClientSugar {
-
-  override def lastSeqAndOffset(
-      targetEventHubNameAndPartition: EventHubNameAndPartition): Option[(Long, Long)] = {
-    val x = eventHubs.messageStore(targetEventHubNameAndPartition).length.toLong - 1
-    Some((x, x))
-  }
-
-  override def lastEnqueuedTime(nameAndPartition: EventHubNameAndPartition): Option[Long] = {
-    Some(
-      eventHubs
-        .messageStore(nameAndPartition)
-        .last
-        .getSystemProperties
-        .getEnqueuedTime
-        .toEpochMilli)
-  }
-
-  override def beginSeqNo(nameAndPartition: EventHubNameAndPartition): Option[Long] = {
-    Some(-1L)
-  }
-}
-
-class TestEventHubsClient(ehParams: Map[String, String],
-                          eventHubs: SimulatedEventHubs,
-                          latestRecords: Map[EventHubNameAndPartition, (Long, Long, Long)])
-    extends Client
-    with TestClientSugar {
-  private var partitionId: Int = _
-  private var offsetType: EventHubsOffsetType = _
-  private var currentOffset: String = _
-
-  override def initReceiver(partitionId: String,
-                            offsetType: EventHubsOffsetType,
-                            currentOffset: String): Unit = {
-    this.partitionId = partitionId.toInt
-    this.offsetType = offsetType
-    this.currentOffset = currentOffset
-  }
-
-  override def receive(expectedEventNum: Int): Iterable[EventData] = {
-    val eventHubName = ehParams("eventhubs.name")
-    if (offsetType != EventHubsOffsetTypes.EnqueueTime) {
-      eventHubs.search(EventHubNameAndPartition(eventHubName, partitionId),
-                       currentOffset.toInt,
-                       expectedEventNum)
-    } else {
-      eventHubs.searchWithTime(EventHubNameAndPartition(eventHubName, partitionId),
-                               ehParams("eventhubs.filter.enqueuetime").toLong,
-                               expectedEventNum)
-    }
-  }
-
-  override def lastSeqAndOffset(
-      nameAndPartition: EventHubNameAndPartition): Option[(Long, Long)] = {
-    val (offset, seq, _) = latestRecords(nameAndPartition)
-    Some(offset, seq)
-  }
-
-  override def lastEnqueuedTime(nameAndPartition: EventHubNameAndPartition): Option[Long] =
-    Some(latestRecords(nameAndPartition)._3)
-
-  override def beginSeqNo(nameAndPartition: EventHubNameAndPartition): Option[Long] =
-    Some(-1L)
-}
-
-class FluctuatedEventHubClient(ehParams: Map[String, String],
-                               eventHubs: SimulatedEventHubs,
-                               ssc: StreamingContext,
-                               messagesBeforeEmpty: Long,
-                               numBatchesBeforeNewData: Int,
-                               latestRecords: Map[EventHubNameAndPartition, (Long, Long)])
-    extends Client
-    with TestClientSugar {
-  private var callIndex = -1
-  private var partitionId: Int = _
-  private var offsetType: EventHubsOffsetType = _
-  private var currentOffset: String = _
-
-  override def initReceiver(partitionId: String,
-                            offsetType: EventHubsOffsetType,
-                            currentOffset: String): Unit = {
-    this.partitionId = partitionId.toInt
-    this.offsetType = offsetType
-    this.currentOffset = currentOffset
-  }
-
-  override def receive(expectedEventNum: Int): Iterable[EventData] = {
-    val eventHubName = ehParams("eventhubs.name")
-    if (offsetType != EventHubsOffsetTypes.EnqueueTime) {
-      eventHubs.search(EventHubNameAndPartition(eventHubName, partitionId),
-                       currentOffset.toInt,
-                       expectedEventNum)
-    } else {
-      eventHubs.searchWithTime(EventHubNameAndPartition(eventHubName, partitionId),
-                               ehParams("eventhubs.filter.enqueuetime").toLong,
-                               expectedEventNum)
-    }
-  }
-
-  override def lastSeqAndOffset(
-      nameAndPartition: EventHubNameAndPartition): Option[(Long, Long)] = {
-    callIndex += 1
-    if (callIndex < numBatchesBeforeNewData) {
-      Some(messagesBeforeEmpty - 1, messagesBeforeEmpty - 1)
-    } else {
-      Some(latestRecords(nameAndPartition))
-    }
-  }
-
-  override def lastEnqueuedTime(nameAndPartition: EventHubNameAndPartition): Option[Long] = {
-    Some(Long.MaxValue)
-  }
-
-  override def beginSeqNo(nameAndPartition: EventHubNameAndPartition): Option[Long] = {
-    Some(-1L)
   }
 }
