+++ core/src/test/scala/org/apache/spark/streaming/eventhubs/EventHubDirectDStreamSuite.scala
@@ -53,7 +53,7 @@ class EventHubDirectDStreamSuite extends EventHubTestSuiteBase with MockitoSugar
     Mockito
       .when(eventHubClientMock.beginSeqNo(Matchers.any[EventHubNameAndPartition]))
       .thenReturn(None)
-    ehDStream.setEventHubClient(ehNameToClient)
+    ehDStream.ehClients = ehNameToClient
     ssc.scheduler.start()
     intercept[IllegalArgumentException] {
       ehDStream.compute(Time(1000))
@@ -75,8 +75,7 @@ class EventHubDirectDStreamSuite extends EventHubTestSuiteBase with MockitoSugar
     Mockito
       .when(eventHubClientMock.lastSeqAndOffset(Matchers.any[EventHubNameAndPartition]))
       .thenReturn(None)
-
-    ehDStream.setEventHubClient(ehNameToClient)
+    ehDStream.ehClients = ehNameToClient
     ssc.scheduler.start()
     intercept[IllegalArgumentException] {
       try {
